More logic (moved to multisubsystemcommands)
