+++ src/main/java/frc/robot/subsystems/drive/Drive.java
@@ -59,13 +59,13 @@ public class Drive extends SubsystemBase {
   private Rotation2d _rawGyroRotation = new Rotation2d();
   private SwerveModulePosition[] _lastModulePositions = // For delta tracking
       new SwerveModulePosition[] {
-        new SwerveModulePosition(),
-        new SwerveModulePosition(),
-        new SwerveModulePosition(),
-        new SwerveModulePosition()
+          new SwerveModulePosition(),
+          new SwerveModulePosition(),
+          new SwerveModulePosition(),
+          new SwerveModulePosition()
       };
-  private SwerveDrivePoseEstimator _poseEstimator =
-      new SwerveDrivePoseEstimator(_kinematics, _rawGyroRotation, _lastModulePositions, new Pose2d());
+  private SwerveDrivePoseEstimator _poseEstimator = new SwerveDrivePoseEstimator(_kinematics, _rawGyroRotation,
+      _lastModulePositions, new Pose2d());
 
   private Alliance _alliance;
   private Pose2d _speakerPosition;
@@ -97,9 +97,8 @@ public class Drive extends SubsystemBase {
         this::runVelocity,
         new HolonomicPathFollowerConfig(
             DriveConstants.MAX_LINEAR_SPEED, DriveConstants.DRIVE_BASE_RADIUS, new ReplanningConfig()),
-        () ->
-            DriverStation.getAlliance().isPresent()
-                && DriverStation.getAlliance().get() == Alliance.Red,
+        () -> DriverStation.getAlliance().isPresent()
+            && DriverStation.getAlliance().get() == Alliance.Red,
         this);
     Pathfinding.setPathfinder(new LocalADStarAK());
     PathPlannerLogging.setLogActivePathCallback(
@@ -113,21 +112,20 @@ public class Drive extends SubsystemBase {
         });
 
     // Configure SysId
-    _sysId =
-        new SysIdRoutine(
-            new SysIdRoutine.Config(
-              DriveConstants.SYSID_RAMP_RRATE,
-              DriveConstants.SYSID_STEP_VOLTAGE,
-              DriveConstants.SYSID_TIMEOUT,
-                (state) -> Logger.recordOutput("Drive/SysIdState", state.toString())),
-            new SysIdRoutine.Mechanism(
-                (voltage) -> {
-                  for (int i = 0; i < 4; i++) {
-                    _modules[i].runCharacterization(voltage.in(Volts));
-                  }
-                },
-                null,
-                this));
+    _sysId = new SysIdRoutine(
+        new SysIdRoutine.Config(
+            DriveConstants.SYSID_RAMP_RRATE,
+            DriveConstants.SYSID_STEP_VOLTAGE,
+            DriveConstants.SYSID_TIMEOUT,
+            (state) -> Logger.recordOutput("Drive/SysIdState", state.toString())),
+        new SysIdRoutine.Mechanism(
+            (voltage) -> {
+              for (int i = 0; i < 4; i++) {
+                _modules[i].runCharacterization(voltage.in(Volts));
+              }
+            },
+            null,
+            this));
     _centerOfRotation = new Translation2d();
     _field = new Field2d();
   }
@@ -138,7 +136,7 @@ public class Drive extends SubsystemBase {
 
     _visionIO.updateInputs(_visionInputs);
     Logger.processInputs("Drive/Vision", _visionInputs);
-        
+
     for (var module : _modules) {
       module.periodic();
     }
@@ -158,11 +156,10 @@ public class Drive extends SubsystemBase {
     SwerveModulePosition[] modulePositions = getModulePositions();
     SwerveModulePosition[] moduleDeltas = new SwerveModulePosition[4];
     for (int moduleIndex = 0; moduleIndex < 4; moduleIndex++) {
-      moduleDeltas[moduleIndex] =
-          new SwerveModulePosition(
-              modulePositions[moduleIndex].distanceMeters
-                  - _lastModulePositions[moduleIndex].distanceMeters,
-              modulePositions[moduleIndex].angle);
+      moduleDeltas[moduleIndex] = new SwerveModulePosition(
+          modulePositions[moduleIndex].distanceMeters
+              - _lastModulePositions[moduleIndex].distanceMeters,
+          modulePositions[moduleIndex].angle);
       _lastModulePositions[moduleIndex] = modulePositions[moduleIndex];
     }
 
@@ -183,13 +180,18 @@ public class Drive extends SubsystemBase {
     }
     _poseEstimator.update(_rawGyroRotation, modulePositions);
 
-    // MITCHELL READ THIS COMMENT. I'm still getting the loop overrun issue from the driver station
-    // so this did not solve the issue. but i wanted to bring it to your attention. 
-    // as of the commit which added this block of comments, the only errors I'm getting from the driver
-    // station is about photon vision cameras and the loop over run from drive periodic. i am a little 
-    // worried that this drive periodic overrun issue might come back to bite us so I don't want to 
-    // leave it not addressed for too long. But, first, lets get the rest of the robot back up and going
-  
+    // MITCHELL READ THIS COMMENT. I'm still getting the loop overrun issue from the
+    // driver station
+    // so this did not solve the issue. but i wanted to bring it to your attention.
+    // as of the commit which added this block of comments, the only errors I'm
+    // getting from the driver
+    // station is about photon vision cameras and the loop over run from drive
+    // periodic. i am a little
+    // worried that this drive periodic overrun issue might come back to bite us so
+    // I don't want to
+    // leave it not addressed for too long. But, first, lets get the rest of the
+    // robot back up and going
+
     // commenting this out to see if it helps with loop overrun time gh - 2/25/24
     // driver.periodic is sometimes running at 20-40ms on its own
     // _field.setRobotPose(_poseEstimator.getEstimatedPosition());
@@ -227,8 +229,10 @@ public class Drive extends SubsystemBase {
   }
 
   /**
-   * Stops the drive and turns the modules to an X arrangement to resist movement. The modules will
-   * return to their normal orientations the next time a nonzero velocity is requested.
+   * Stops the drive and turns the modules to an X arrangement to resist movement.
+   * The modules will
+   * return to their normal orientations the next time a nonzero velocity is
+   * requested.
    */
   public void stopWithX() {
     Rotation2d[] headings = new Rotation2d[4];
@@ -249,7 +253,10 @@ public class Drive extends SubsystemBase {
     return _sysId.dynamic(direction);
   }
 
-  /** Returns the module states (turn angles and drive velocities) for all of the modules. */
+  /**
+   * Returns the module states (turn angles and drive velocities) for all of the
+   * modules.
+   */
   @AutoLogOutput(key = "SwerveStates/Measured")
   private SwerveModuleState[] getModuleStates() {
     SwerveModuleState[] states = new SwerveModuleState[4];
@@ -259,9 +266,13 @@ public class Drive extends SubsystemBase {
     return states;
   }
 
-  /** Returns the module positions (turn angles and drive velocities) for all of the modules. */
+  /**
+   * Returns the module positions (turn angles and drive velocities) for all of
+   * the modules.
+   */
   // TODO: Garrett 2/27/24
-  // should this return _module[i].getModueState() instead? Advantagekit cannot show a SwerveModulePosition. 
+  // should this return _module[i].getModueState() instead? Advantagekit cannot
+  // show a SwerveModulePosition.
   // only a swerve module state.
   @AutoLogOutput(key = "SwerveStates/MeasuredPosition")
   private SwerveModulePosition[] getModulePositions() {
@@ -296,7 +307,7 @@ public class Drive extends SubsystemBase {
    * Adds a vision measurement to the pose estimator.
    *
    * @param visionPose The pose of the robot as measured by the vision camera.
-   * @param timestamp The timestamp of the vision measurement in seconds.
+   * @param timestamp  The timestamp of the vision measurement in seconds.
    */
   public void addVisionMeasurement(Pose2d visionPose, double timestamp) {
     _poseEstimator.addVisionMeasurement(visionPose, timestamp);
@@ -315,17 +326,29 @@ public class Drive extends SubsystemBase {
   private Pose2d getSpeakerPos() {
     if (_speakerPosition == null) {
       if (getAlliance() != null) {
-        _speakerPosition = (getAlliance() == DriverStation.Alliance.Blue) ? DriveConstants.BLUE_SPEAKER : DriveConstants.RED_SPEAKER;
+        _speakerPosition = (getAlliance() == DriverStation.Alliance.Blue) ? DriveConstants.BLUE_SPEAKER
+            : DriveConstants.RED_SPEAKER;
       }
     }
 
     return _speakerPosition;
   }
 
-  /** Returns the distance from the center of the robot to the alliance's speaker */
+  /**
+   * Returns the distance from the center of the robot to the alliance's speaker
+   */
   public double getRadiusToSpeakerInMeters() {
-    
+
     return getRadiusToSpeakerInMeters(_poseEstimator.getEstimatedPosition(), getSpeakerPos());
+  }
+  
+  /**
+   * Returns the distance from the center of the robot to the alliance's speaker.
+   * NOTE: Use THIS constructor for GrpShootOnTheMove only.
+   */
+  public double getRadiusToSpeakerInMeters(Pose2d robotPose) {
+
+    return getRadiusToSpeakerInMeters(robotPose, getSpeakerPos());
   }
 
   public SwerveDrivePoseEstimator getPoseEstimator() {
@@ -344,7 +367,8 @@ public class Drive extends SubsystemBase {
     return this;
   }
 
-  // this setup lets us test the math, but when we actually run the code we don't have to give a pose estimator
+  // this setup lets us test the math, but when we actually run the code we don't
+  // have to give a pose estimator
   public static double getRadiusToSpeakerInMeters(Pose2d robotPose, Pose2d speakerPos) {
     double xDiff = robotPose.getX() - speakerPos.getX();
     double yDiff = robotPose.getY() - speakerPos.getY();
@@ -371,7 +395,11 @@ public class Drive extends SubsystemBase {
   }
 
   /**
-   * NOTE: Only used for CmdShootDriveOnTheMove; otherwise, use calcAngleToSpeaker.
+   * Thank you to Team 3467 - Windham Windup!
+   * Adopted from 3467's getAngleToSpeaker:
+   * https://github.com/WHS-FRC-3467/Skip-5.14-Nocturne/blob/PostGSDCleanup/src/main/java/frc/robot/Util/FieldCentricAiming.java
+   * NOTE: Only used for GrpShootDriveOnTheMove; otherwise, use
+   * calcAngleToSpeaker.
    */
 
   public Rotation2d getRotation2dToSpeaker(Translation2d futurePose) {
@@ -411,10 +439,10 @@ public class Drive extends SubsystemBase {
   /** Returns an array of module translations. */
   public static Translation2d[] getModuleTranslations() {
     return new Translation2d[] {
-      new Translation2d(DriveConstants.TRACK_WIDTH_X / 2.0, DriveConstants.TRACK_WIDTH_Y / 2.0),
-      new Translation2d(DriveConstants.TRACK_WIDTH_X / 2.0, -DriveConstants.TRACK_WIDTH_Y / 2.0),
-      new Translation2d(-DriveConstants.TRACK_WIDTH_X / 2.0, DriveConstants.TRACK_WIDTH_Y / 2.0),
-      new Translation2d(-DriveConstants.TRACK_WIDTH_X / 2.0, -DriveConstants.TRACK_WIDTH_Y / 2.0)
+        new Translation2d(DriveConstants.TRACK_WIDTH_X / 2.0, DriveConstants.TRACK_WIDTH_Y / 2.0),
+        new Translation2d(DriveConstants.TRACK_WIDTH_X / 2.0, -DriveConstants.TRACK_WIDTH_Y / 2.0),
+        new Translation2d(-DriveConstants.TRACK_WIDTH_X / 2.0, DriveConstants.TRACK_WIDTH_Y / 2.0),
+        new Translation2d(-DriveConstants.TRACK_WIDTH_X / 2.0, -DriveConstants.TRACK_WIDTH_Y / 2.0)
     };
   }
 }
